move base, exc, log to utils, minor fixes
