+++ pyLinguist/utils/exc.py
@@ -1,3 +1,5 @@
+""" Used exceptions """
+
 import logging
 
 
@@ -16,12 +18,11 @@ ERROR_CODES = {
 
 class YaTranslateException(Exception):
     """ Yandex API exceptions """
-    error_codes = ERROR_CODES
 
     def __init__(self, status_code: int, *args, **kwargs):
         if not isinstance(status_code, int):
             logging.warning("int required, got %s", type(status_code))
-        message = self.error_codes.get(
+        message = ERROR_CODES.get(
             status_code, "Unknown code {}".format(status_code)
         )
         super(YaTranslateException, self).__init__(
